feat: add lesson endpoint
