+++ apps/api/src/lesson/lesson.module.ts
@@ -2,15 +2,16 @@ import { Module } from "@nestjs/common";
 
 import { FileModule } from "src/file/files.module";
 
-import { AdminLessonRepository } from "./adminLesson.repository";
-import { AdminLessonService } from "./adminLesson.service";
+import { AdminLessonRepository } from "./repositories/adminLesson.repository";
+import { AdminLessonService } from "./services/adminLesson.service";
 import { LessonController } from "./lesson.controller";
-import { LessonRepository } from "./lesson.repository";
+import { LessonRepository } from "./repositories/lesson.repository";
+import { LessonService } from "./services/lesson.service";
 
 @Module({
   imports: [FileModule],
   controllers: [LessonController],
-  providers: [LessonRepository, AdminLessonService, AdminLessonRepository],
+  providers: [LessonRepository, AdminLessonService, AdminLessonRepository, LessonService],
   exports: [AdminLessonService, AdminLessonRepository, LessonRepository],
 })
 export class LessonModule {}
