+++ apps/api/src/seed/seed.ts
@@ -163,7 +163,7 @@ async function createLessonProgress(userId: string) {
       studentId: userId,
       createdAt: courseLesson.createdAt,
       updatedAt: courseLesson.createdAt,
-      quizScore: courseLesson.lessonType === LESSON_TYPES.quiz ? 0 : null,
+      quizScore: courseLesson.lessonType === LESSON_TYPES.QUIZ ? 0 : null,
     };
   });
 
@@ -191,7 +191,7 @@ async function createQuizAttempts(userId: string) {
     .innerJoin(chapters, eq(courses.id, chapters.courseId))
     .innerJoin(lessons, eq(lessons.chapterId, chapters.id))
     .innerJoin(questions, eq(questions.lessonId, lessons.id))
-    .where(and(eq(courses.isPublished, true), eq(lessons.type, LESSON_TYPES.quiz)))
+    .where(and(eq(courses.isPublished, true), eq(lessons.type, LESSON_TYPES.QUIZ)))
     .groupBy(courses.id, lessons.id);
 
   const createdQuizAttempts = quizzes.map((quiz) => {
